+++ flutter_reach_five/example/lib/methods/update_profile_method.dart
@@ -96,7 +96,7 @@ class UpdateProfileMethodState extends State<UpdateProfileMethod> {
               consentSupportedKey: Consent(
                 date: DateTime.now(),
                 granted: consentGranted,
-              )
+              ),
           },
         );
       });