+++ flutter_reach_five/test/helpers/reach_five_key_converter_test.dart
@@ -35,7 +35,7 @@ void main() {
             providerCreators: [
               ProviderCreatorInterface(
                 type: ProviderCreatorTypeInterface.google,
-              )
+              ),
             ],
           ),
         ),