+++ reach_five_identity_repo/lib/src/api/o_auth_api.dart
@@ -28,7 +28,7 @@ class OAuthApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future]
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<void>> revokeToken({
     required RevokeTokenRequest revokeTokenRequest,
     CancelToken? cancelToken,
@@ -57,12 +57,12 @@ class OAuthApi {
     try {
       _bodyData = jsonEncode(revokeTokenRequest);
     } catch (error, stackTrace) {
-      throw DioError(
+      throw DioException(
         requestOptions: _options.compose(
           _dio.options,
           _path,
         ),
-        type: DioErrorType.unknown,
+        type: DioExceptionType.unknown,
         error: error,
         stackTrace: stackTrace,
       );